pyproject.toml
[project]
name = "sentiment-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "altair>=6.0.0",
    "datasets>=4.1.1",
    "emoji>=2.15.0",
    "llama-cpp-python>=0.3.0",
    "marimo>=0.23.1",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "pandas>=3.0.2",
    "plotly>=6.7.0",
    "ruff>=0.15.10",
    "seaborn>=0.13.2",
    "torch>=2.11.0",
    "transformers>=4.57.0",
    "ty>=0.0.29",
]
